requirements.txt
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
cryptography==38.0.3
future==0.18.2
fuzzywuzzy==0.18.0
gkeepapi==0.14.2
gpsoauth==1.0.2
idna==3.4
importlib-metadata==5.0.0
jaraco.classes==3.2.3
jeepney==0.8.0
keyring==23.11.0
more-itertools==9.0.0
oauthlib==3.2.2
pycparser==2.21
pycryptodomex==3.15.0
python-dateutil==2.8.2
python3-discogs-client==2.5
requests==2.28.1
SecretStorage==3.3.3
six==1.16.0
titlecase==2.4
urllib3==1.26.12
zipp==3.10.0
